refactor(benchttp): split files
