+++ src/benchttp/common.ts
@@ -1,0 +1,1 @@
+export type GoDuration = `${number}${'ns' | 'µs' | 'ms' | 's' | 'm' | 'h'}`
