+++ src/benchttp/tests.ts
@@ -1,0 +1,1 @@
+export type TestPredicate = 'EQ' | 'NEQ' | 'GT' | 'GTE' | 'LT' | 'LTE'
